+++ src/OdfAttributeName.ts
@@ -1,9 +1,18 @@
 export enum OdfAttributeName {
+  FormatBreakBefore = "fo:break-before",
+  FormatColor = "fo:color",
+  FormatFontSize = "fo:font-size",
+  FormatFontStyle = "fo:font-style",
+  FormatFontWeight = "fo:font-weight",
+  FormatTextAlign = "fo:text-align",
+
   OfficeMimetype = "office:mimetype",
   OfficeVersion = "office:version",
 
   StyleFamily = "style:family",
   StyleName = "style:name",
+  StylePosition = "style:position",
+  StyleType = "style:type",
 
   TextAnchorType = "text:anchor-type",
   TextOutlineLevel = "text:outline-level",