+++ src/index.ts
@@ -4,10 +4,13 @@ export { TextDocument } from "./TextDocument";
 export { Image } from "./draw/Image";
 
 // style
+export { Color } from "./style/Color";
 export { HorizontalAlignment } from "./style/HorizontalAlignment";
-export { Style } from "./style/Style";
+export { IParagraphStyle } from "./style/IParagraphStyle";
+export { ParagraphStyle } from "./style/ParagraphStyle";
 export { TabStop } from "./style/TabStop";
 export { TabStopType } from "./style/TabStopType";
+export { Typeface } from "./style/Typeface";
 
 // text
 export { Heading } from "./text/Heading";