+++ src/style/Typeface.ts
@@ -1,0 +1,8 @@
+export enum Typeface {
+  Normal,
+  Bold,
+  Italic,
+  Oblique,
+  BoldItalic,
+  BoldOblique,
+}