+++ test/integration.spec.ts
@@ -1,15 +1,23 @@
 import { unlink } from "fs";
 import { join } from "path";
 import { promisify } from "util";
+import { Color } from "../src/style/Color";
 import { HorizontalAlignment } from "../src/style/HorizontalAlignment";
-import { Style } from "../src/style/Style";
+import { ParagraphStyle } from "../src/style/ParagraphStyle";
 import { TabStop } from "../src/style/TabStop";
 import { TabStopType } from "../src/style/TabStopType";
+import { Typeface } from "../src/style/Typeface";
 import { TextDocument } from "../src/TextDocument";
 
 const FILEPATH = "./integration.fodt";
 
-describe("integration", () => {
+xdescribe("integration", () => {
+  let document: TextDocument;
+
+  beforeAll(() => {
+    document = new TextDocument();
+  });
+
   afterAll(async (done) => {
     const unlinkAsync = promisify(unlink);
 
@@ -18,42 +26,86 @@ describe("integration", () => {
     done();
   });
 
-  it("create a full blown document", async (done) => {
-    const document = new TextDocument();
-
+  it("image", () => {
     document.addParagraph().addImage(join(__dirname, "data", "ODF.png"));
+  });
 
+  it("add heading", () => {
     document.addHeading("First heading");
     document.addHeading("Second heading", 2);
 
-    const para1 = document.addParagraph("The quick, brown fox jumps over a lazy dog.");
-    para1.addText("\nSome more text");
-    para1.setStyle(new Style());
-    para1.getStyle().setHorizontalAlignment(HorizontalAlignment.Center);
+    const para = document.addParagraph("The quick, brown fox jumps over a lazy dog.");
+    para.addText("\nSome more text");
+  });
 
-    document.addParagraph();
+  describe("paragraph formatting", () => {
+    it("page break", () => {
+      const heading = document.addHeading("Paragraph Formatting", 2);
+      heading.setStyle(new ParagraphStyle());
+      heading.getStyle().setPageBreakBefore();
+    });
 
-    const para2 = document.addParagraph("first\tsecond\tthird");
-    para2.setStyle(new Style());
-    para2.getStyle().addTabStop(new TabStop(4));
-    para2.getStyle().addTabStop(new TabStop(12, TabStopType.Right));
+    it("align text", () => {
+      const paragraph = document.addParagraph("Some centered text");
+      paragraph.setStyle(new ParagraphStyle());
+      paragraph.getStyle().setHorizontalAlignment(HorizontalAlignment.Center);
+    });
 
-    const heading20 = document.addHeading("List");
-    heading20.setStyle(new Style());
-    heading20.getStyle().setPageBreakBefore();
+    it("tab stops", () => {
+      const paragraph = document.addParagraph("first\tsecond\tthird");
+      paragraph.setStyle(new ParagraphStyle());
+      paragraph.getStyle().addTabStop(new TabStop(4));
+      paragraph.getStyle().addTabStop(new TabStop(12, TabStopType.Right));
+    });
+  });
+
+  describe("text formatting", () => {
+    beforeAll(() => {
+      const heading = document.addHeading("Text Formatting", 2);
+      heading.setStyle(new ParagraphStyle());
+      heading.getStyle().setPageBreakBefore();
+    });
+
+    it("color", () => {
+      const paragraph = document.addParagraph("Some mint-colored text");
+      paragraph.setStyle(new ParagraphStyle());
+      paragraph.getStyle().setColor(Color.fromRgb(62, 180, 137));
+    });
+
+    it("font size", () => {
+      const paragraph = document.addParagraph("Some small text");
+      paragraph.setStyle(new ParagraphStyle());
+      paragraph.getStyle().setFontSize(8);
+    });
+
+    it("typeface", () => {
+      const paragraph = document.addParagraph("Some bold text");
+      paragraph.setStyle(new ParagraphStyle());
+      paragraph.getStyle().setTypeface(Typeface.Bold);
+    });
+  });
+
+  it("hyperlink", () => {
+    const heading = document.addHeading("Hyperlink", 2);
+    heading.setStyle(new ParagraphStyle());
+    heading.getStyle().setPageBreakBefore();
+
+    const paragraph = document.addParagraph("This is just an ");
+    paragraph.addHyperlink("example", "http://example.org");
+    paragraph.addText(".");
+  });
+
+  it("list", () => {
+    const heading = document.addHeading("List", 2);
+    heading.setStyle(new ParagraphStyle());
+    heading.getStyle().setPageBreakBefore();
 
     const list = document.addList();
     list.addItem("first item");
     list.addItem("second item");
+  });
 
-    const heading30 = document.addHeading("Another chapter");
-    heading30.setStyle(new Style());
-    heading30.getStyle().setPageBreakBefore();
-
-    const para3 = document.addParagraph("This is just an ");
-    para3.addHyperlink("example", "http://example.org");
-    para3.addText(".");
-
+  it("save document", async (done) => {
     await document.saveFlat(FILEPATH);
 
     // TODO use snapshot testing